requirements.txt
mysql-connector-python
yowsup==3.2.3
tqdm
rfc6266-parser
six==1.10
urllib3
requests
pyaxmlparser
eventlet
consonance==0.1.3.post1
wget
transitions==0.6.9
bcrypt==3.1.7
python-dateutil==2.5.2
pycrypto==2.6.1
devops-tools
cryptography

